requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
# SQLite is included in Python standard library 